+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Released]
 
+## 2022-03-30 12:00:00 
+
+### Added
+- Added job for expiring the conversion
+
 ## 2022-03-25 12:00:00 
 
 ### Added